add check on gym status
added additional check on whether gym is open before displaying and completing certain actions
handles cases where gyms are closed during usual hours due to external reasons (e.g. public holidays)

current logic is to check if first gym is open or closed, and assumes that this determines the rest of the gyms.
logic will have to be changed in the future to check for each individual gym, and display based on its status